requirements.txt
pandas
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

aioredis==2.0.1
redis==4.6.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.0

httpx==0.25.2
aiohttp==3.9.1


python-magic==0.4.27
pillow==10.1.0
pytesseract==0.3.10
PyPDF2==3.0.1
python-docx==1.1.0
chardet==5.2.0
aiofiles==23.2.1


markdown==3.5.1
pymdown-extensions==10.5


python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0.post1


loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.39.1


pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2


black==23.12.0
flake8==6.1.0
isort==5.13.0
mypy==1.7.1


celery[redis]==5.3.4
flower==2.0.1


sse-starlette==1.8.2


python-dotenv==1.0.0


fastapi-cors==0.0.6


slowapi==0.1.9


pydantic[email]==2.5.0

tzdata==2023.3


orjson==3.9.10


gunicorn==21.2.0

gitpython==3.1.40         
websockets==12.0          
rich==13.7.0              
click==8.1.7              
watchfiles==0.21.0        
docker==7.0.0             
psutil==5.9.6             